Add post with metadata
The post needs to contain the meta information to be parsed correctly.

Since the post will be published in the English categories it is unclear
whether Softwareentwicklung or a English category should be supplied.

The tags only contain NuGet to begin with but could be expanded. Not
sure yet which tags would fit the article.

Signed-off-by: Scherb Vincent <vincent.scherb@adesso.de>